+++ HISTORY.md
@@ -38,6 +38,9 @@
     * German
     * Chinese, thanks, Simon!
 
+### 2021-10-08
+* objects: tweaked variable block visibility
+
 ### 2021-10-07
 * objects, byob: new BlockVisibilityDialogMorph for bulk-selecting blocks to hide / show in the palette
 * ojects: simplified palette context menu